# [inductor][Autotune] Add matrix_instr_nonkdim to triton_meta